Move IRI parsing and encoding to its own class
